+++ js/const.js
@@ -1,0 +1,15 @@
+/* eslint-disable no-unused-vars */
+
+const IMAGE_LIST_LENGTH = 25;
+
+const LIKES_COUNT = {
+  min: 15,
+  max: 200,
+};
+
+const COMMENTS_COUNT = {
+  min: 0,
+  max: 30,
+};
+
+export {IMAGE_LIST_LENGTH, LIKES_COUNT, COMMENTS_COUNT};
